fix: Output full first context line, including indentation
Previously, we would omit any and all content that precedes the
match, on the first line of the match context. In particular, if
the first line of a match has initial whitespace, then this would
be trimmed in the output.

Unfortunately, in order to reconstruct the first line fully, we
must re-read it from the originating file. This is done using a
line scanner, to avoid reading the entire file into memory.
Nevertheless, this is sub-optimal; in the future, it may be
worthwhile improving on this implementation.

Also: show only a single-line code extract, when the policy
dictates that parent context be omitted.