Rename rules
The format of the rules description was done in a way that didn't explicitly mention what was the problem,  but what was the overall expectation. Considering we use it in the CLI output with associated snippet code in errors, it makes more sense to format it in a way that explain what is the problem.